+ Added the prototype level with lighting
The lighting I have used for the time being is the directional lights as for some reason I couldn't get any others to work.

They will serve their purpose for the prototype but will need to be changed. If anyone would like to change them for the prototype, please do.

I have also changed the doorways to ones that are more easy to implement, if there is time I can come back and add the doors as shown in the level plan.