! Improved script readabillity
I have renamed all scripts so they are easier to read and reference.

To show who made a script, if anyone made any changes and a description of what the script does, I added comments at the top of each script. If any changes are made to a script please update these comments and add them when creating a new script.